Add a base class to write document based fix all providers
The base class reduces code dublication and provides a clean way to write a fix all provider that fixes each document separatly. Also use this class in our fix all providers.
