+++ StyleCop.Analyzers/StyleCop.Analyzers/ReadabilityRules/SA1107FixAllProvider.cs
@@ -1,64 +1,18 @@
 ﻿namespace StyleCop.Analyzers.ReadabilityRules
 {
-    using System.Collections.Generic;
     using System.Collections.Immutable;
     using System.Threading.Tasks;
     using Microsoft.CodeAnalysis;
-    using Microsoft.CodeAnalysis.CodeActions;
     using Microsoft.CodeAnalysis.CodeFixes;
     using Microsoft.CodeAnalysis.CSharp;
     using Microsoft.CodeAnalysis.Editing;
+    using StyleCop.Analyzers.Helpers;
 
-    internal class SA1107FixAllProvider : FixAllProvider
+    internal class SA1107FixAllProvider : DocumentBasedFixAllProvider
     {
-        public override async Task<CodeAction> GetFixAsync(FixAllContext fixAllContext)
-        {
-            switch (fixAllContext.Scope)
-            {
-            case FixAllScope.Document:
-                var newRoot = await this.FixAllInDocumentAsync(fixAllContext, fixAllContext.Document).ConfigureAwait(false);
-                return CodeAction.Create(ReadabilityResources.SA1107CodeFix, token => Task.FromResult(fixAllContext.Document.WithSyntaxRoot(newRoot)));
+        protected override string CodeActionTitle => ReadabilityResources.SA1107CodeFix;
 
-            case FixAllScope.Project:
-                Solution solution = await this.GetProjectFixesAsync(fixAllContext, fixAllContext.Project).ConfigureAwait(false);
-                return CodeAction.Create(ReadabilityResources.SA1107CodeFix, token => Task.FromResult(solution));
-
-            case FixAllScope.Solution:
-                var newSolution = fixAllContext.Solution;
-                var projectIds = newSolution.ProjectIds;
-                for (int i = 0; i < projectIds.Count; i++)
-                {
-                    newSolution = await this.GetProjectFixesAsync(fixAllContext, newSolution.GetProject(projectIds[i])).ConfigureAwait(false);
-                }
-
-                return CodeAction.Create(ReadabilityResources.SA1107CodeFix, token => Task.FromResult(newSolution));
-
-            case FixAllScope.Custom:
-            default:
-                return null;
-            }
-        }
-
-        private async Task<Solution> GetProjectFixesAsync(FixAllContext fixAllContext, Project project)
-        {
-            Solution solution = project.Solution;
-            var oldDocuments = project.Documents.ToImmutableArray();
-            List<Task<SyntaxNode>> newDocuments = new List<Task<SyntaxNode>>(oldDocuments.Length);
-            foreach (var document in oldDocuments)
-            {
-                newDocuments.Add(this.FixAllInDocumentAsync(fixAllContext, document));
-            }
-
-            for (int i = 0; i < oldDocuments.Length; i++)
-            {
-                var newDocumentRoot = await newDocuments[i].ConfigureAwait(false);
-                solution = solution.WithDocumentSyntaxRoot(oldDocuments[i].Id, newDocumentRoot);
-            }
-
-            return solution;
-        }
-
-        private async Task<SyntaxNode> FixAllInDocumentAsync(FixAllContext fixAllContext, Document document)
+        protected override async Task<SyntaxNode> FixAllInDocumentAsync(FixAllContext fixAllContext, Document document)
         {
             DocumentEditor editor = await DocumentEditor.CreateAsync(document, fixAllContext.CancellationToken).ConfigureAwait(false);
