+++ StyleCop.Analyzers/StyleCop.Analyzers/StyleCop.Analyzers.csproj
@@ -200,6 +200,7 @@
     <Compile Include="MaintainabilityRules\SA1406DebugFailMustProvideMessageText.cs" />
     <Compile Include="MaintainabilityRules\SA1407ArithmeticExpressionsMustDeclarePrecedence.cs" />
     <Compile Include="MaintainabilityRules\SA1407SA1408CodeFixProvider.cs" />
+    <Compile Include="Helpers\DocumentBasedFixAllProvider.cs" />
     <Compile Include="MaintainabilityRules\SA1407SA1408FixAllProvider.cs" />
     <Compile Include="MaintainabilityRules\SA1408ConditionalExpressionsMustDeclarePrecedence.cs" />
     <Compile Include="MaintainabilityRules\SA1409RemoveUnnecessaryCode.cs" />
